+++ app/ExternalFiles/Extensions/Real_Import.php
@@ -12,11 +12,19 @@ defined( 'ABSPATH' ) || exit;
 
 use ExternalFilesInMediaLibrary\Dependencies\easySettingsForWordPress\Settings;
 use ExternalFilesInMediaLibrary\ExternalFiles\Extension_Base;
+use ExternalFilesInMediaLibrary\ExternalFiles\ImportDialog;
 
 /**
  * Handler controls how to import external files for real in media library without external connection.
  */
 class Real_Import extends Extension_Base {
+	/**
+	 * The internal extension name.
+	 *
+	 * @var string
+	 */
+	protected string $name = 'real_import';
+
 	/**
 	 * Instance of actual object.
 	 *
@@ -69,6 +77,15 @@ class Real_Import extends Extension_Base {
 	}
 
 	/**
+	 * Return the object title.
+	 *
+	 * @return string
+	 */
+	public function get_title(): string {
+		return __( 'Really import each file', 'external-files-in-media-library' );
+	}
+
+	/**
 	 * Add our custom settings for this plugin.
 	 *
 	 * @return void
@@ -78,7 +95,7 @@ class Real_Import extends Extension_Base {
 		$settings_obj = Settings::get_instance();
 
 		// get the advanced section.
-		$advanced_tab_advanced = $settings_obj->get_section( 'settings_section_advanced' );
+		$advanced_tab_advanced = $settings_obj->get_section( 'settings_section_dialog' );
 
 		// bail if section could not be loaded.
 		if ( ! $advanced_tab_advanced ) {
@@ -86,13 +103,13 @@ class Real_Import extends Extension_Base {
 		}
 
 		// add setting.
-		$setting = $settings_obj->add_setting( 'eml_directory_listing_real_import' );
+		$setting = $settings_obj->add_setting( 'eml_real_import' );
 		$setting->set_section( $advanced_tab_advanced );
 		$setting->set_field(
 			array(
 				'type'        => 'Checkbox',
-				'title'       => __( 'Really import each file', 'external-files-in-media-library' ),
-				'description' => __( 'If this option is enabled each external URL will be imported as real file in your media library. They will not be "external files" in your media library. This setting overrides user-specific settings if enabled.', 'external-files-in-media-library' ),
+				'title'       => $this->get_title(),
+				'description' => __( 'If this option is enabled each external URL will be imported as real file in your media library. They will not be "external files" in your media library.', 'external-files-in-media-library' ),
 			)
 		);
 		$setting->set_type( 'integer' );
@@ -173,12 +190,17 @@ class Real_Import extends Extension_Base {
 	 * @return array<string,mixed>
 	 */
 	public function add_option_in_form( array $dialog ): array {
+		// only add if it is enabled in settings.
+		if( ! in_array( $this->get_name(), ImportDialog::get_instance()->get_enabled_extensions(), true ) ) {
+			return $dialog;
+		}
+
 		// get the actual state for the checkbox.
 		$checked = 1 === absint( get_option( 'eml_real_import' ) );
 
 		// if user has its own setting, use this.
-		if ( 1 === absint( get_user_meta( get_current_user_id(), 'efml_real_import', true ) ) ) {
-			$checked = true;
+		if ( ImportDialog::get_instance()->is_customization_allowed() ) {
+			$checked = 1 === absint( get_user_meta( get_current_user_id(), 'efml_real_import', true ) );
 		}
 
 		// collect the entry.